refactor: Refactor Auth module
Relocate constants from high-level to module level. Add an Http status codes for auth.controller
