+++ src/modules/auth/auth.service.ts
@@ -7,7 +7,7 @@ import {
 import { PrismaService } from '../prisma/prisma.service';
 import { RegisterDto } from './dto/register.dto';
 import { AuthRepository } from './auth.repository';
-import { Constants } from '../../common/constants';
+import { Constants } from './constants';
 import * as bcrypt from 'bcrypt';
 import { LoginDto } from './dto/login.dto';
 import { JwtTokensService } from './jwt.token.service';
